:recycle: Refactor: refactor backtest option code split 백테스트 옵션 styles & types 관련 코드 분리
